feat:combined apis into one handler
