+++ department/function.yml
@@ -1,26 +1,16 @@
-addDepartment:
-    handler: department/addDepartment.handler
+department:
+    handler: department/department.handler
     role: ${env:LAMBDA_ROLE}
     events:
         - http:
               path: /department
               method: post
-            #   cors: true
-
-getDepartment:
-    handler: department/getDepartment.handler
-    role: ${env:LAMBDA_ROLE}
-    events:
+              cors: true
         - http:
               path: /department
               method: get
-            #   cors: true
-
-updateDepartment:
-    handler: department/updateDepartment.handler
-    role: ${env:LAMBDA_ROLE}
-    events:
+              cors: true
         - http:
               path: /department
               method: put
-            #   cors: true                            
+              cors: true
