+++ empType/function.yml
@@ -1,26 +1,17 @@
-addEmptype:
-    handler: empType/addEmpType.handler
-    role: ${env:LAMBDA_ROLE}
-    events:
-        - http:
-              path: /empType
-              method: post
-              cors: true
-
-getEmptype:
-    handler: empType/getEmpType.handler
-    role: ${env:LAMBDA_ROLE}
-    events:
-        - http:
-              path: /empType
-              method: get
-              cors: true 
-
-updateEmptype:
-    handler: empType/updateEmpType.handler
-    role: ${env:LAMBDA_ROLE}
-    events:
-        - http:
-              path: /empType
-              method: put
-              cors: true                                          
+empType:
+  handler: empType/empType.handler
+  role: ${env:LAMBDA_ROLE}
+  events:
+    - http:
+        path: /empType
+        method: post
+        cors: true
+    - http:
+        path: /empType
+        method: get
+        cors: true
+    - http:
+        path: /empType
+        method: put
+        cors: true
+
